Simplified the creation of the output file path.
This construction:
```
output_path = Path(phrase_file.with_suffix(".mp3").name)
```

Replaces the phrase_file suffix, ".txt", with ".mp3" and then creates
a truncated relative Path. If `phrase_file` is an absolute path, the
previous implementation would create the output file in whatever
directory the input file was located in. This will create the output
file in the current working directory.

It mostly makes testing easier and is less surprising to the user.