Updated tests to use fixtures
The tests/conftest.py holds fixtures that can be used in the various
test functions. The directory test/resources contains a `phrases.txt`
file and a `mustang4.wav` for use in testing the operation of the
AudioProgramGenerator class.

Updated `test_apg.py` to include tests for creating a
AudioProgramGenerator with good and bad inputs.

Created `test_main.py` to test the AudioProgramGenerator.apg.main
function which uses docopt to define and parse arguments. Arguments
are spoofed by setting `sys.argv` to the arguments to test.