Created a new command-line interface based on typer
Changes made to support new typer-based implementation:
- added `typer` to installation requirements
- added `ncli` to [tools.poetry.scripts] section in pyproject.toml
- created audio_program_generator/__main__.py to house new CLI

The new CLI uses [typer][1] to specify the arguments handled
by the cli as well as docstrings which will appear when an
error occurs while parsing the arguments supplied by the user
as well as when the user asks for help with `--help`.

The outline of a typer-based implementation is first create a
`typer.Typer()` instance which will stand in for our `main()`
function. Next, attach functions to the `Typer` instance using
decorators (i.e. @cli.command), and the options to that function
will specify it's command-line options.

1: https://typer.tiangolo.com